# Perf: Collect trivias - remove unnecessary sorts