before
| Method |    Mean |   Error |  StdDev | Rank |        Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|------- |--------:|--------:|--------:|-----:|-------------:|------------:|-----------:|----------:|
| Format | 35.87 s | 4.575 s | 0.251 s |    1 | 2150000.0000 | 663000.0000 | 80000.0000 |   10.7 GB |

after
| Method |    Mean |   Error |  StdDev | Rank |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|------- |--------:|--------:|--------:|-----:|------------:|------------:|-----------:|----------:|
| Format | 16.48 s | 7.260 s | 0.398 s |    1 | 665000.0000 | 247000.0000 | 35000.0000 |   3.47 GB |

Not bad for 3 lines remove 😛